requirements.txt
python-telegram-bot==20.7
firebase-admin==6.5.0
google-generativeai==0.8.3
bcrypt==4.2.0
python-dotenv==1.0.1
asyncio==3.4.3